chat/pyproject.toml
[project]
name = "chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opentelemetry-distro==0.43b0",
    "streamlit>=1.28.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.1.0",
    "python-dotenv>=1.0.0",
    "pip>=25.2",
]
